Create README.md
I have a wide range of interests and with family and projects I tend to stay busy, that's just how I like it. I work closely with an Information Systems department and a huge strength of mine is the ability to recognize patterns. I would like to collaborate with someone who is far more tech savy than I am. I am willing to go 50-50 with someone to ensure we have a viable product without any need for hurt feelings or greed. I will trust you until given reason not to and seek the same.
Thanks and have a good one.